pyproject.toml
[tool.poetry]
name = "aether-blender"
version = "0.1.0"
description = "Natural language interface for Blender 3D animation. Chat with AI to create professional animations without learning Blender."
authors = ["BytedotGit"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/BytedotGit/aether-blender"
keywords = ["blender", "ai", "animation", "3d", "natural-language"]
package-mode = false
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.dependencies]
python = "^3.11"
# AI Providers
anthropic = "^0.18.0"
openai = "^1.12.0"
# GUI
PyQt6 = "^6.6.0"
# Utilities
pydantic = "^2.6.0"
python-dotenv = "^1.0.0"
# Blender type stubs (for autocomplete outside Blender)
fake-bpy-module = "^20240927"
google-generativeai = "^0.8.6"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
pytest-qt = "^4.3.0"
# Code Quality
black = "^24.1.0"
ruff = "^0.1.14"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
# Development
ipython = "^8.21.0"

[tool.poetry.scripts]
aether = "src.orchestrator.launcher:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | tools
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument
]
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "tools",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG001", "ARG002", "S101"]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
exclude = [
    "tools/",
    ".venv/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "e2e: end-to-end tests requiring Blender",
    "slow: slow running tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "src/blender_addon/*",  # Runs inside Blender, not testable with pytest
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
